Optimized the code of Myfooter and added animation effects to the pictures of sponsors.
Using v-for instead of the lis, decreased the number of code, and use ref array to loading the href and img src in every lis.
This will be more maintainable.(But, note that every index of imgs must be corresponding with the index of links)

Animation use gsap may lead to need more time in the first loading of the page. But now we are more attention at how to make the awesome effect, so we will ignore it now. And, when we come to the optimized turn, we may discuss about it again.(This will only record one time. It will not appear in the next time I use gsap in this project.)

We give the pictures a fade-in effect with 20px Y offest and 1s duration. Each lis will be delayed by 0.2 seconds before starting the animation.