Added animation effect to p tag in Myfooter
Now we use the transition method of Vue to make the animation.
We originally wanted to use the chain call of GSAP, but because the effect is OK, and it may take some time to debug the delay parameter after the change, so we gave up and use it in the next animation effect.