Some updates have been made and vuex module has been installed
Explored the use of vuex for global state management and event processing bus.

The animation for nav had been made for the home page, and explored the message passing of the components in the route to the global nav.

Basically realized the appearance and disappearance effect of NAV, but there is still a bug.

(When you enter the home page for the first time or refresh the home page, NAV will appear first and then disappear.)

Although the impact is small, it is actually difficult for me to solve this problem. The essential reason is that the rendering function updates so fast that the default value must be read first every time it is refreshed. Therefore, NAV appears by default on all pages, but home is not. However, there can only be one default value, and a special default value cannot be given to home. I try to modify the default value, However, the rendering function executes too fast. I start rendering every time before I modify the data. Moreover, if I sleep the rendering function, it will affect the loading performance of the whole page, which is completely unacceptable.

Now, my only solution is to use watch to monitor the modification and play the skeleton interface before the modification is completed. However, this is also difficult. Such changes are too large and the effect needs to be evaluated. At present, I don't have a good solution to monitor only the home page. If other pages also need to monitor, the amount of code will increase and become more difficult to maintain, And the changes will become bigger. Maybe this is not a good way. Maybe I will have a better solution in the future, but now, let it stay there.

I'll create a new branch to tell me that I can explore the results of using native Vue instead of vuex (which is very simple). Of course, we don't study it now. Just like multipage, it just gives a possibility for the future.