Remove ref sugars
They are unstable experimental properties.

And their types need to be explicitly referred is a very troublesome thing.

Especially when it is an experimental feature and may have to be modified, which means that it may not be explicitly referred in the future. Adding a new file will cause code redundancy and affect performance.