Final record
In the end, the problem cannot be solved. When using the vite packaged program (JS import resources in the asset) on the iGEM wiki, the files in the project file need to be used, but the iGEM official website provides HTML files. To obtain JS files, you must initiate a get request to the PHP backend of the official website for analysis, so as to return the correct JS file content without suffix, HTTP links cannot be used for import in rollup packaged JS. Moreover, even if the browser supports it (DeNO has been implemented in the planning, and chrome is unknown), rollup itself is searched in the same folder after packaging, and they cannot be disassembled and renamed. In essence, they are bundled, but disassembled for reuse, It also reduces the amount of data in a single request. In order to optimize the experience, it doesn't work in the current iGEM wiki mode. There may be a solution to package it into a single JS file, but I tried it and failed. I need to continue to study the mechanism of rollup and the source code that may involve some Vites providing information to rollup (probably not involved), But now I have no time for in-depth research. Wikis must be completed quickly and at a high speed, so it is impossible to continue to study rollup settings. Moreover, even if I fully understand it, there may be no good way to complete the packaging of single files. In the end, there may be no gain or loss, and high investment may have no return. Then I will really die, and the project may not be completed, That's a big trouble for me. I don't allow such a thing to happen, so I gave up this scheme, used a template with CDN like introduction to complete the development of a single page, and then completed all pages one by one) so this project was abandoned

A new project will be launched, using the single HTML page mode and introducing in-line JS and CSS to realize the use of templates and rapid visual development. It will be tested directly on the iGEM server, reduce the redundancy of intermediate steps, and get feedback and repair quickly

The packages that will be migrated in the past include vue3, GSAP and maybe tailwind. The packages that may be introduced in the future include bootstrap